+++ tests/main.spec.js
@@ -52,14 +52,14 @@ describe('Base tests', () => {
   });
 
   it('Subscribe method returns function', () => {
-    unsubscribe = store.$$subscribe((value) => value);
+    unsubscribe = store.$$instance.subscribe((value) => value);
     expect(typeof unsubscribe).toBe('function');
     unsubscribe();
   });
 
   it('Subscribe listener emits value on update', (done) => {
     let iterIdx = 0;
-    unsubscribe = store.$$subscribe((value) => {
+    unsubscribe = store.$$instance.subscribe((value) => {
       if (!iterIdx) {
         iterIdx++;
         return;